Scan recursively to figure out if project is using yarn
