+++ packages/core/src/packager/index.ts
@@ -1,0 +1,1 @@
+export * as Packager from "./packager";
